requirements.txt
langchain-community
langchain-google-genai
langchain
streamlit
python-dotenv
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
python-magic-bin
sentence-transformers